Pass proper dirty rect in paint event to QOpenGLWidget with partial updates
When QOpenGLWidget has partial updates enabled, there is no reason to
render entire scene on every frame. Instead, the widget may rerender
only a portion of it. All the data has been cached by Qt in the internal
framebuffer object. This limiting is necessary when rendering a complex
scene on 4k display. See https://bugs.kde.org/show_bug.cgi?id=413504

The patch adds a special flag into QWidgetPrivate to notify
QWidgetBackingStore that this widget needs proper tracking of
the dirty region.

Theoretically, flushing of the textures could also be limited to
the dirty region, but it is a different task.