Fix unbalanced KeyPress/Release events in children of QMdiArea
When the user presses Ctrl+Tab, QMdiArea is supposed to switch
the active child window. When doing so, it eats the event. The
problem is that is doesn't eat the ShortcutOverride event,
which is kept unbalanced with the absent KeyRelease event.

See https://bugs.kde.org/show_bug.cgi?id=424319