Fix shortcuts assigned to special keys on non-latin layouts
When the user has non-latin layout we still consider latin
meaning of the keys to be a priorty. This behavior is already
in Qt (and is disputable), but this patch doesn't change that.
The patch only adds proper handling of special keys:

* '\' --- backslash/bar key
* '/' --- slash/question key
* ';' --- semicolon/colon key
* '`' --- backqoute/tilde key
* ''' --- quote/doublequote key
* '[' --- left bracket key
* ']' --- right bracket key
* ',' --- comma/< key
* '.' --- period/> key
* '-' --- minus/underscore key
* '=' --- equal/plus key

See https://bugs.kde.org/show_bug.cgi?id=430479