[win+angle] Fix flickering when resizing the Window
When the user resizes the window, DXGI surface becomes outdated. We need
to call eglWaitNative to make sure the surface gets in sync with the
window size. The D3D implementation in ANGLE doesn't actually wait for anything
in this call. It just compares the size of the surface and adjusts it if necessary.

Co-authored-by: Alvin Wong <alvinhochun@gmail.com>