Add QMdiArea::AlwaysShowSubwindowNameInTitleBar
This option makes Qt always put the active window name into the
title bar of the main window, even in tabbar or floating window
mode.

This patch is **extremely** hackish, because the original design
of the title update in Qt in weird. There is are model/view and
all the children of QMdiArea manage the parent's title in a
"distributed" (read, random) way.

See this bug: https://bugs.kde.org/show_bug.cgi?id=475172