[win] Generate Enter/Leave events on touching non-client area of the window
When the **tablet** pointer enters the non-client area of the window, we should emit
the leave event. And we should also emit an Enter event when the pointer re-enters
the window from its non-client area.

Failing to do so leaves the cursor in a broken state (it sticks to a "resize" cursor).

Steps to reproduce (with a tablet device):

1) From the outside of the window, slowly enter the non-client area and stop
when the resizing arrows appear

2) Slowly enter the client area of the window until the resizeing arrows disappear

3) Slowly enter the non-client area again to see the arrows. **DO NOT** leave the windows,
   i.e. don't let the arrows disappear.

4) Enter the client area again. Now the cursor is stuck to the resizing arrows.