Fix slowdown caused by making window system events synchronous.
The slowdown is very noticeable on Android most probably because of the
way Qt dispatches input events. When done one thread, Android has to
wait before let's say the redraw is complete. And then because the
subsequent event tends to be almost the same, it does redraw again. But
when we do this on two separate threads, Qt tends to batch them and only
redraw when absolutely necessary, making things considerably faster.

cc: @dkazakov