Implement better parsing of EDID model number
When extended display sections are not available for these values,
the algorithm falls back to the values from the header:
QScreen::model() is assigned with
"<Manifacturer ID><Manufacturer product code (hex)>", and
QScreen::serialNumber() is assigned with "<Serial number(hex)>".

The patch also ports the manual little-endian parsing to
qFromLittleEndian.