Implement color conversion for the backing store texture
If the window surface is not in sRGB mode, then the backing store
surface should be converted into the destinations color space.

This patch adds the most popular color space transitions into
QBackingStoreDefaultCompositor.

The patch also implements QOpenGLWidget::setTextureColorSpace(),
which notifies the compositor about the color space of non-native
openGL widgets, so that the data could be converted correctly.

TODO: should we implement the same for QOpenGLWindow::
setTextureColorSpace()?

Note:
The channels should be swizzled into RGBA *before* applying the
color space conversion matrix. Otherwise the colors will be skewed,
because the conversion function for R and B channels is not the same.