[win] Recover ANGLE support on Windows
The angle support has been removed from 6.x branch. This patch
recovers this functionality and fixes a few issues in
nativeplatform API implementation.

Raw openGL implementation on Windows hugely depends on the drivers
provided by the vendors. And oftentimes it is not good enough.

More than that, raw openGL API does not allow the usage of HDR
functionality, which is available in DXGI-only.

The patch also does the following changes:

1) Recovers QWindowsEGLStaticContext, which implements API
   for accessing openGLES via ANGLE

2) Connects the ANGLE backend to the feature QT_FEATURE_egl.
   The feature is enabled automatically, when EGL library is found

3) Updates FindEGL.cmake module from ECM 6.7.0

4) Removes QT_FEATURE_dynamicgl. This feature doesn't make
   sense anymore. On Windows openGL implementation is **always**
   dynamic, without any options.

   (in some places the patch temporarily replaces it with
    QT_FEATURE_opengl_dynamic, which will also be removed in
	a later patch)

This reverts commit 752497910b67b2a1a80560840ca44548d8893434.