Implement a manual test for checking is HDR features work
Test plan:

1) Run without arguments: `hdr-openglwidget.exe`
   It should show you three rectangles: the left one should be HDR'ly
   bright, the other ones should be SDR'ly dim and look exactly the same.

3) Run in Bt. 2020 PQ mode: `hdr-openglwidget.exe --bt2020pq`
   The result should look exactly the same.

4) Run in SDR sRGB mode: `hdr-openglwidget.exe --srgb`.
   All three images should look SDR'ly dim.

NOTE:
Please note that the current implementation of SDR compositing
in QOpenGLTextureBlitter doesn't support user configuration for
SDR brightness from the system. This API is available for UWP
applications only. It means that when changing "SDR brightness"
slider in Windows' settings, the brightness of our SDR widget
will not change. More that that, it might even be different from
the brightness of other SDR applications.

Change-Id: Idccc790937c9061ec618ab21f6b71bd0620cd2cc