Add fast-transfer option for QImage when rendering QWidget
Copying the image for every frame is really expensive, especially
on HiDPI screen. It may eat up to 15ms, which prevents the app
from rendering with 60 FPS rate.

This lazy-copying might be unsafe for some of QRhi backends,
which would render in a separate thread, but I couldn't find any
traces of it.