Compress frames in a more parallel way
The patch changes the frame we check for completion. Now we
don't check the "last" frame, but "one but last". The problem is
that QWidgetRepaintManager also performs a lot of CPU-intensive
operations during the frame cycle, so to make sure it has time to do
that we should allow piling one frame in the rendering queue.

The patch adds QRhi::isOneButLastFrameCompletedOnGPU() to acheve
this functionality.