Add workaround for flickering on resizing on ANGLE and Qt6
The patch implements a set of hacks that (almost) fixes
black stripes artifacts and flickering on resizing when
using ANGLE on Windows as a backend.

It seems like on Windows, the compositor starts rendering
the surface much earlier than WM_SIZE is arrived to the app.
(I guess it considers the window as "resized" right after the
delivery of WM_SIZING, but I have no proof of that).
That is, ideally, by the arival of WM_SIZE our rendering chain
should already be resized and updated. Othrewise, the
compositor tries to render the old version of the surface's
texture on screen, which created visual artifacts.

Another problem is that ANGLE internally uses GetClientRect()
WINAPI call to calculate the "preferred" surface size. And this size
returns the "old" size until the final WM_SIZE arrives, which means
in the time frame between WM_SIZING and WM_SIZE ANGLE
still uses the "old" size, even though the compositor already renders
it on a bigger window.

The patch addresses these issues separately by allowing to activate
workarounds by setting specific environment variables:

1) `set QT_USE_PREMATURE_RESIZE_EVENTS=1` moves the delivery of
    QResizeEvent from WM_SIZE to WM_SIZING. It means that the new
    frame is requested much earlier and it **may be** ready by the time the
    compositor actually renders our window. I have to use "may be", becasue
    the actual frame rendering might be delayed by the events system and
    the framerate compressor. But we have no proper way of requesting
    the update synchronously.

    The sequence of events before requestign the frame looks like that:

    1) Geometry of the widget is changed
    2) QLayout::widgetEvent() receives either QEvent::LayoutRequest or
        QEvent::Resize.
    3) The layouted widgets are resized and then request updates individually

    The whole sequence happens asynchronously via events queue. So I have
    no idea how to linearize it into a sequential synchronous process.

2) `set QT_ANGLE_MANUALLY_UPDATE_SURFACE_SIZE=1` forces ANGLE's
    EGL_ANGLE_window_fixed_size into action. It basically lets the app to control
    the surface size manually. Coupled with QT_USE_PREMATURE_RESIZE_EVENTS
    it forces ANGLE to resize the internal surface right in WM_SIZING event, without
    waiting for WM_SIZE and GetClientRect() to get a proper value.

When enabled together, they almost remove flickering and black stripes problems
during resizing Krita window in non-native canvas mode. When enabling native
canvas mode, the artifacts return back. I'm not sure why. I think it happens because
of the frames compressor and it just skips the frames/updates for one of the surfaces.
Though I'm not sure if that is true.

CC:kimageshop@kde.org