Fix flickering when opening dialogs on ANGLE and Qt6
It seems like Qt has a bit of incorrect(?) way of displaying the windows when using
the QRhi method. Qt first sets the window as visible, creates the swap chain, and only
then starts all the resize, relayout and render passes.

It means that the compositor (and DXGI) starts to rendering the widnow's surface right
away (even before the first Present1() call). It causes all the windows and dialogs
momentarily flicker with black/white color before displaying anything sane.

Ideally, it would be better to postpone the creation of the swap chain until the window
passes through all the resize/relayout/render stages (i.e. WM_SIZE, WM_ERASEBKGND and
WM_PAINT). Instead we create the swap chain in QWidgetPrivate::create(), way before
the window is ready to be shown to the user.

This patch does **not** fix the bug properly. Instead it adds a workaround and prefills
the newly created surface with the window backgroud color. This way flickering is not much
noticable.

To activate the workaround, do `set QT_PREFILL_RHI_SURFACE=1` and it will prefill the
surface 1) on creation; 2) on every resize that happens before the first WM_PAINT.

PS:
There is still another issue that the window may be briefly shown at a random position
(and possibly on a different screen) and then moved to a proper location. That is how
window initialization happens on Windows. We should just avoid creation of the swap
chain that early.

PPS:
The patch will not work if the surface uses non-sRGB color space, since no conversion is done.

CC:kimageshop@kde.org